# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Audisor Assistant API
  version: 0.1.0
  description: >
    Authenticated backend API for the Audisor Writing & Design Assistant.
    Covers operation lifecycle management, operator chat with tool-calling,
    A-Flow issue management, and token usage estimation.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/operations:
    post:
      tags: [operations]
      summary: Create a new operation
      operationId: create_operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationRequest"
      responses:
        "200":
          description: Operation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
        "422":
          description: Validation error
  /v1/operations/{operation_id}/status:
    get:
      tags: [operations]
      summary: Get operation status
      operationId: operation_status
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Current operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
        "404":
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
  /v1/operations/{operation_id}/cancel:
    post:
      tags: [operations]
      summary: Cancel an operation
      operationId: cancel_operation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
        "404":
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
  /v1/operations/{operation_id}/resume:
    post:
      tags: [operations]
      summary: Resume a suspended operation
      operationId: resume_operation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeOperationRequest"
      responses:
        "200":
          description: Operation resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
        "404":
          description: Operation not found
        "409":
          description: Conflict (e.g. version mismatch or invalid resume state)
  /v1/operations/{operation_id}/claim:
    post:
      tags: [operations]
      summary: Atomically claim a pending tool execution
      description: >
        Binds a suspended operation's tool call to a specific claimant (tab).
        First valid claim wins; subsequent claims from the same claimant are
        idempotent; claims from different claimants are rejected with 409.
      operationId: claim_tool_execution
      parameters:
        - $ref: "#/components/parameters/OperationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimToolExecutionRequest"
      responses:
        "200":
          description: Claim accepted or idempotent replay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimToolExecutionResponse"
        "404":
          description: Operation not found
        "409":
          description: Competing claim or conflict
  /v1/operations/{operation_id}/events:
    get:
      tags: [operations]
      summary: Read operation events after a cursor
      operationId: operation_events
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - name: after
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Sequence number to read after
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum events to return
      responses:
        "200":
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationEventsResponse"
        "404":
          description: Operation not found

  # ─── Assistant ───────────────────────────────────────────────────────────────
...
  /v1/aflow/issues:
    get:
      tags: [aflow-management]
      summary: List A-Flow issues
      operationId: aflow_issues
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum issues to return
      responses:
        "200":
          description: Paginated issue list
          content:
            application/json:
              schema:
                type: object
  /v1/aflow/issues/{issue_id}:
    get:
      tags: [aflow-management]
      summary: Get a single A-Flow issue
      operationId: aflow_issue
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Issue not found
  /v1/aflow/issues/{issue_id}/retry-local:
    post:
      tags: [aflow-management]
      summary: Retry an issue against the local provider
      operationId: retry_local
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryRequest"
      responses:
        "200":
          description: Retry initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
        "404":
          description: Issue not found
        "409":
          description: Retry not supported or conflict
  /v1/aflow/issues/{issue_id}/continue-with-fallback:
    post:
      tags: [aflow-management]
      summary: Retry an issue against the fallback provider (Fireworks)
      operationId: continue_with_fallback
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryRequest"
      responses:
        "200":
          description: Fallback retry initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
        "404":
          description: Issue not found
        "409":
          description: Fallback not ready or conflict

  # ─── Usage ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Authentication is enforced through the AuthProvider boundary.
        Production vendor selection is pending.
  parameters:
    OperationId:
      name: operation_id
      in: path
      required: true
      schema:
        type: string
      description: Unique operation identifier
  schemas:
...
    CreateOperationRequest:
      type: object
      additionalProperties: false
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        source_kind:
          type: string
          enum: [task, prepared_plan, fix]
          default: task
        plan:
          type: object
          nullable: true
          description: Pre-computed plan (for prepared_plan source_kind)
    ResumeOperationRequest:
      type: object
      additionalProperties: false
      required: [resume_type, suspension_id, operation_version, payload]
      properties:
        resume_type:
          type: string
        suspension_id:
          type: string
        operation_version:
          type: integer
        payload:
          type: object
    ClaimToolExecutionRequest:
      type: object
      additionalProperties: false
      required: [claimant_id, suspension_id, call_id, argument_digest]
      properties:
        claimant_id:
          type: string
          description: Unique tab/claimant identifier
        suspension_id:
          type: string
          description: Suspension to claim
        call_id:
          type: string
          description: Tool call identifier
        argument_digest:
          type: string
          description: SHA-256 digest of call_id + tool_name + canonical JSON arguments
    ClaimToolExecutionResponse:
      type: object
      additionalProperties: false
      required:
        - operation_id
        - state
        - claimed
        - claimant_id
        - call_id
        - tool_name
        - arguments
        - argument_digest
        - suspension_id
      properties:
        operation_id:
          type: string
        state:
          type: string
        claimed:
          type: boolean
        claimant_id:
          type: string
        call_id:
          type: string
        tool_name:
          type: string
        arguments:
          type: object
        argument_digest:
          type: string
        suspension_id:
          type: string
        idempotent:
          type: boolean
          default: false
    OperationStatusResponse:
      type: object
      properties:
        operation_id:
          type: string
        state:
          type: string
        detail:
          type: object
    OperationEventsResponse:
      type: object
      properties:
        operation_id:
          type: string
        events:
          type: array
          items:
            type: object
        cursor:
          type: integer
        terminal:
          type: boolean
    RetryRequest:
      type: object
      additionalProperties: false
      required: [operation_version, idempotency_key]
      properties:
        operation_version:
          type: integer
          minimum: 0
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 128

    # ─── Assistant ───────────────────────────────────────────────────────────
...
    HTTPError:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: object